+++ lib/activerecord-multi-tenant.rb
@@ -1,6 +1,8 @@
+require 'activerecord-multi-tenant/controller_extensions'
 require 'activerecord-multi-tenant/copy_from_client'
 require 'activerecord-multi-tenant/default_scope'
 require 'activerecord-multi-tenant/migrations'
+require 'activerecord-multi-tenant/model_extensions'
 require 'activerecord-multi-tenant/multi_tenant'
 require 'activerecord-multi-tenant/referential_integrity'
 require 'activerecord-multi-tenant/version'