+++ lib/activerecord-multi-tenant/default_scope.rb
@@ -3,7 +3,7 @@ class ActiveRecord::Base
     alias :unscoped_orig :unscoped
     def unscoped
       if respond_to?(:scoped_by_tenant?) && MultiTenant.current_tenant_id
-        unscoped_orig.where(arel_table[MultiTenant.partition_key].eq(MultiTenant.current_tenant_id))
+        unscoped_orig.where(arel_table[self.partition_key].eq(MultiTenant.current_tenant_id))
       else
         unscoped_orig
       end