+++ spec/activerecord-multi-tenant/model_extensions_spec.rb
@@ -1,0 +1,153 @@
+require 'spec_helper'
+
+describe MultiTenant do
+  after { MultiTenant.current_tenant = nil }
+
+  # Setting and getting
+  describe 'Setting the current tenant' do
+    before { MultiTenant.current_tenant = :foo }
+    it { MultiTenant.current_tenant == :foo }
+  end
+
+  describe 'is_scoped_as_tenant should return the correct value when true' do
+    it {expect(Project.respond_to?(:scoped_by_tenant?)).to eq(true)}
+  end
+
+  describe 'is_scoped_as_tenant should return the correct value when false' do
+    it {expect(UnscopedModel.respond_to?(:scoped_by_tenant?)).to eq(false)}
+  end
+
+  context 'immutability' do
+    before do
+      @account = Account.create! name: 'foo'
+      @project = @account.projects.create! name: 'bar'
+    end
+
+    describe 'tenant_id should be immutable, if already set' do
+      it { expect { @project.account_id = @account.id + 1 }.to raise_error(MultiTenant::TenantIsImmutable) }
+    end
+
+    describe 'setting tenant_id to the same value should not error' do
+      it { expect { @project.account_id = @account.id }.not_to raise_error }
+    end
+
+    describe 'setting tenant_id to a string with same to_i value should not error' do
+      it { expect { @project.account_id = @account.id.to_s }.not_to raise_error }
+    end
+  end
+
+  describe 'tenant_id should auto populate after initialization' do
+    before do
+      @account = Account.create! name: 'foo'
+      MultiTenant.current_tenant = @account
+    end
+    it {expect(Project.new.account_id).to eq(@account.id)}
+  end
+
+  describe 'Handles custom partition_key on tenant model' do
+    before do
+      @account  = Account.create! name: 'foo'
+      MultiTenant.current_tenant = @account
+      @custom_partition_key_task = CustomPartitionKeyTask.create! name: 'foo'
+    end
+
+    it { expect(@custom_partition_key_task.account).to eq(@account) }
+  end
+
+  # Scoping models
+  describe 'Project.all should be scoped to the current tenant if set' do
+    before do
+      @account1 = Account.create! name: 'foo'
+      @account2 = Account.create! name: 'bar'
+
+      @project1 = @account1.projects.create! name: 'foobar'
+      @project2 = @account2.projects.create! name: 'baz'
+
+      MultiTenant.current_tenant = @account1
+      @projects = Project.all
+    end
+
+    it { expect(@projects.length).to eq(1) }
+    it { expect(@projects).to eq([@project1]) }
+  end
+
+  describe 'Querying the tenant from a scoped model with a tenant set' do
+    before do
+      @account = Account.create! name: 'foo'
+      @project = @account.projects.create! name: 'foobar'
+      MultiTenant.current_tenant= @account1
+    end
+
+    it { @project.account }
+  end
+
+  # Associations
+  describe 'Associations should be correctly scoped by current tenant' do
+    before do
+      @account = Account.create! name: 'foo'
+      @project = Project.create! name: 'foobar', account: @account
+
+      MultiTenant.current_tenant = @account
+      @task = @project.tasks.create! name: 'baz'
+    end
+
+    it 'should correctly set the tenant on the task created with current_tenant set' do
+      expect(@task.account).to eq(@account)
+    end
+  end
+
+  describe "It should be possible to use aliased associations" do
+    before do
+      @account = Account.create! name: 'baz'
+      MultiTenant.current_tenant = @account
+    end
+
+    it { expect(AliasedTask.create(:name => 'foo', :project_alias => @project2).valid?).to eq(true) }
+  end
+
+  describe "It should be possible to use associations with partition_key from polymorphic" do
+    before do
+      @account = Account.create!(name: 'foo')
+      MultiTenant.current_tenant = @account
+      @project = Project.create!(name: 'project', account: @account)
+      @comment = Comment.new commentable: @project, account: @account
+    end
+
+    it { expect(@comment.save!).to eq(true) }
+  end
+
+  # ::with
+  describe "::with" do
+    it "should set current_tenant to the specified tenant inside the block" do
+      @account = Account.create!(:name => 'baz')
+
+      MultiTenant.with(@account) do
+        expect(MultiTenant.current_tenant).to eq(@account)
+      end
+    end
+
+    it "should reset current_tenant to the previous tenant once exiting the block" do
+      @account1 = Account.create!(:name => 'foo')
+      @account2 = Account.create!(:name => 'bar')
+
+      MultiTenant.current_tenant = @account1
+      MultiTenant.with @account2 do
+
+      end
+
+      expect(MultiTenant.current_tenant).to eq(@account1)
+    end
+
+    it "should return the value of the block" do
+      @account1 = Account.create!(:name => 'foo')
+      @account2 = Account.create!(:name => 'bar')
+
+      MultiTenant.current_tenant = @account1
+      value = MultiTenant.with @account2 do
+        "something"
+      end
+
+      expect(value).to eq "something"
+    end
+  end
+end