+++ spec/activerecord-multi-tenant/record_finding_spec.rb
@@ -2,10 +2,10 @@ require 'spec_helper'
 
 describe MultiTenant, 'Record finding' do
   it 'searches for tenant object using the scope' do
-    tenant = Tenant.create! name: 'test'
-    obj = tenant.tenant_objects.create! title: 'something'
-    MultiTenant.with(tenant) do
-      expect(TenantObject.find(obj.id)).to be_present
+    account = Account.create! name: 'test'
+    project = account.projects.create! name: 'something'
+    MultiTenant.with(account) do
+      expect(Project.find(project.id)).to be_present
     end
   end
 end