+++ spec/activerecord-multi-tenant/record_modifications_spec.rb
@@ -1,20 +1,20 @@
 require 'spec_helper'
 
 describe MultiTenant, 'Record modifications' do
-  let(:tenant) { Tenant.create! name: 'test' }
-  let(:obj) { tenant.tenant_objects.create! title: 'something' }
+  let(:account) { Account.create! name: 'test' }
+  let(:project) { account.projects.create! name: 'something' }
 
   it 'includes the tenant_id in UPDATEs' do
-    obj.update! title: 'something else'
-    MultiTenant.with(tenant) do
-      expect(TenantObject.find(obj.id).title).to eq 'something else'
+    project.update! name: 'something else'
+    MultiTenant.with(account) do
+      expect(Project.find(project.id).name).to eq 'something else'
     end
   end
 
   it 'includes the tenant_id in DELETEs' do
-    obj.destroy!
-    MultiTenant.with(tenant) do
-      expect(TenantObject.find_by(id: obj.id)).not_to be_present
+    project.destroy!
+    MultiTenant.with(account) do
+      expect(Project.find_by(id: project.id)).not_to be_present
     end
   end
 end