+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.3.0       2016-12-30
+
+* Remove dependency on acts_as_tenant - instead copy the code thats necessary
+* Fix issue with callbacks having TenantIdWrapper instead of the actual object
+
+
 ## 0.2.1       2016-12-29
 
 * Fix bug in CopyFromClient helper